dual osc per key finally removed, 61 osc only, AVR 2.0

+++ ESP/src/serial0.h
@@ -13,7 +13,7 @@
 
 void serial0_init();
 
-void serial0_write_number(uint_fast8_t number, uint_fast8_t newline);
+void serial0_write_number(int_fast32_t number, uint_fast8_t newline);
 
 void serial0_write_hex(uint_fast32_t number);
